BE/requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
djangorestframework_simplejwt
psycopg2-binary==2.8.4